[refactor] re-use exciting sub-functions for plotting spectra in Spectrum class
