+++ pypython/spectrum/_spectrum/plot.py
@@ -10,6 +10,7 @@ import matplotlib.pyplot as plt
 
 from pypython.utilities import plot
 from pypython.spectrum._spectrum import util
+from pypython.spectrum._spectrum import plot_util
 
 
 class SpectrumPlot(util.SpectrumUtil):
@@ -19,23 +20,19 @@ class SpectrumPlot(util.SpectrumUtil):
         """Initialize the class."""
         super().__init__(root, directory, **kwargs)
 
-    def plot_extracted_spectrum(self, thing, fig=None, ax=None, figsize=(10, 6), axes_scales="loglog"):
+    def plot_extracted_spectrum(self, thing, fig=None, ax=None, axes_scales="loglog"):
         """Plot a spectrum, from the spectral cycles."""
 
         if not fig and not ax:
-            fig, ax = plt.subplots(figsize=figsize)
+            fig, ax = plt.subplots(1, 1, figsize=(12, 5))
         elif not fig and ax or fig and not ax:
             raise ValueError("fig and ax need to be supplied together")
 
-        spectral_axis = self["spec"]["Lambda"]
-        flux = self["spec"][thing]
+        ax = plot_util.plot_spectrum(ax, self, thing, None, None, 1.0, "loglog", False)
 
-        ax.plot(spectral_axis, flux, label=f"{thing}" + r"$^{\circ}$")
-        ax.set_xlabel("Spectral axis")
-        ax.set_ylabel("Flux")
-        ax.legend()
+        # if label_lines:
+        #     ax = add_line_ids(ax, common_lines(spectrum=spectrum), "none")
 
-        ax = plot.set_axes_scales(ax, axes_scales)
         fig = plot.finish_figure(fig, "spectrum")
 
         return fig, ax
